Fixed data conversion errors
Fixed data conversion errors
Changed how libraries and includes are linked to a project

+++ Coursework/element_sim.cpp
@@ -130,7 +130,12 @@ void element_sim::render(const glm::vec2 mouse_position) const
 	   );
 	}
 
-    level_->circle_bres(mouse_position.x, mouse_position.y, brush_radius_, renderer_);
+    level_->circle_bres(
+        static_cast<int>(mouse_position.x), 
+        static_cast<int>(mouse_position.y), 
+        brush_radius_, 
+        renderer_
+    );
 
 }
