+++ Coursework/text_renderer.cpp
@@ -87,7 +87,7 @@ void text_renderer::load(const std::string& font, unsigned int font_size)
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            static_cast<unsigned int>(face->glyph->advance.x)
         };
         characters.insert(std::pair<char, character>(c, character_buffer));
     }
